% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df.R
\name{clean_df}
\alias{clean_df}
\title{Reformats raw data frame into a clean data frame}
\usage{
clean_df(df)
}
\arguments{
\item{df}{A raw data frame.}
}
\value{
A clean data frame.
}
\description{
Reformats raw data frame into a clean data frame
}
\examples{

\dontrun{
year <- 2022
html <- fetch_html_element(year)
df_raw <- html_to_df(html)
df_clean <- clean_df(df_raw)
df_clean
}



}
