% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_range_to_dates.R
\name{date_range_to_dates}
\alias{date_range_to_dates}
\title{Converts date range to individual dates}
\usage{
date_range_to_dates(cell, year)
}
\arguments{
\item{cell}{A date range. E.g. "21.02. - 01.03.". As this function is
meant for processing holiday data, those date ranges most
likely stem from data frame cells (E.g. holidays_2020_dates).}

\item{year}{The year from which the dates are from.}
}
\value{
A string where all individual dates are concatenated. Note that
while input date ranges must be in dd.mm. format, the output uses yyyy-mm-dd.
}
\description{
Converts date range to individual dates
}
\examples{

\dontrun{
# Load dataset

data(holidays_2020_dates)
data(holidays_2025_dates)


# Grab different formats of date string

case_1 <- holidays_2020_dates[1,3]  # "-"
case_2 <- holidays_2020_dates[3,3]  # "21.02. - 01.03."
case_3 <- holidays_2020_dates[5,4]  # "04.04. - 19.04.+22.05."
case_4 <- holidays_2020_dates[7,4]  # "25.03.+04.04. - 18.04.+22.05. - 23.05."
case_5 <- holidays_2020_dates[13,4] # "02.04. - 19.04.+04.05.+22.05."
case_6 <- holidays_2020_dates[30,4] # 30.03. - 13.04.+21.05. - 01.06."
case_7 <- holidays_2020_dates[18,7] # "07.12.+19.12. - 03.01."
case_8 <- holidays_2025_dates[6,5]  # "07.07."
case_9 <- holidays_2025_dates[24,5] # "09.06.+19.06.+05.07. - 17.08."


# Resulting date strings

date_range_to_dates(case_1, 2020) # "-"
date_range_to_dates(case_2, 2020) # "2020-02-21, 2020-02-22, ..., 2020-03-01"
# ...
date_range_to_dates(case_9, 2020) # "2020-06-09, 2020-06-19, ..., 2020-08-17"
}



}
