% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_html_element.R
\name{fetch_html_element}
\alias{fetch_html_element}
\title{Fetches html table containing holiday data from website}
\usage{
fetch_html_element(year)
}
\arguments{
\item{year}{The year you want the data from.}
}
\value{
An html table element.
}
\description{
Fetches html table containing holiday data from website
}
\examples{
year <- 2022
fetch_html_element(year)



}
