% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_holidays_2023_dates.R
\docType{data}
\name{holidays_2023_dates}
\alias{holidays_2023_dates}
\title{Swiss school holidays (2023)}
\format{
A data frame with 34 rows and 7 variables:
\describe{
\item{Kanton}{The Swiss canton}
\item{Schule}{The particular type of school (e.g. Alle Stufen, Gymnasium, ...)}
\item{Sportferien}{Date when Sportferien take place}
\item{Fruehlingsferien}{Date when Fruehlingsferien take place}
\item{Sommerferien}{Date when Sommerferien take place}
\item{Herbstferien}{Date when Herbstferien take place}
\item{Weihnachtsferien}{Date when Weihnachtsferien take place}
}
}
\source{
\url{https://www.schulferien.org/schweiz/ferien/}
}
\usage{
holidays_2023_dates
}
\description{
A dataset containing school holidays (1.- 9. grade) as date ranges in the year 2023.
}
\keyword{datasets}
