% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_holidays_20XX_dates.R
\docType{data}
\name{holidays_2019_dates}
\alias{holidays_2019_dates}
\alias{holidays_2020_dates}
\alias{holidays_2021_dates}
\alias{holidays_2022_dates}
\alias{holidays_2023_dates}
\alias{holidays_2024_dates}
\alias{holidays_2025_dates}
\title{Dataset - Swiss school holidays overview}
\format{
Data frames with region and different types of holiday variables.
\describe{
\item{Kanton}{The Swiss canton}
\item{Schule}{The particular type of school (e.g. Alle Stufen, Gymnasium, ...)}
\item{Sportferien}{Date when Sportferien take place}
\item{Fruehlingsferien}{Date when Fruehlingsferien take place}
\item{Sommerferien}{Date when Sommerferien take place}
\item{Herbstferien}{Date when Herbstferien take place}
\item{Weihnachtsferien}{Date when Weihnachtsferien take place}
}

An object of class \code{data.frame} with 36 rows and 7 columns.

An object of class \code{data.frame} with 36 rows and 7 columns.

An object of class \code{data.frame} with 36 rows and 7 columns.

An object of class \code{data.frame} with 36 rows and 7 columns.

An object of class \code{data.frame} with 36 rows and 7 columns.

An object of class \code{data.frame} with 36 rows and 7 columns.
}
\source{
\url{https://www.schulferien.org/schweiz/ferien/}
}
\usage{
holidays_2019_dates

holidays_2020_dates

holidays_2021_dates

holidays_2022_dates

holidays_2023_dates

holidays_2024_dates

holidays_2025_dates
}
\description{
Datasets containing school holidays (1.- 9. grade) as date ranges from the
years 2019 - 2025. This data was scraped in July 2022.
}
\keyword{datasets}
