% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_holidays_20XX_long.R
\docType{data}
\name{holidays_2020_long}
\alias{holidays_2020_long}
\alias{holidays_2021_long}
\alias{holidays_2022_long}
\alias{holidays_2023_long}
\alias{holidays_2024_long}
\alias{holidays_2025_long}
\title{Dataset - Swiss school holidays for ML (Scope: cantons)}
\format{
Data frames with region, temporal, and holiday variables.
\describe{
\item{Kanton}{Swiss cantons, e.g. Aargau, Bern.}
\item{Schule}{The particular type of school, e.g. Alle Stufen, Gymnasium.}
\item{Grossregion}{Major regions of Switzerland, e.g. Zentralschweiz, Ostschweiz.}
\item{Datum}{Date in ISO 8601 format: YYYY-MM-DD.}
\item{Tag}{Day of the month. A value in the range of 01 - 31.}
\item{Monat}{Month of the year. A value in the range of 01 - 12.}
\item{Jahr}{The year, e.g. 2020, 2021.}
\item{Jahreszeit}{Season of the year, e.g. Fruehling, Sommer.}
\item{Ferientag}{Binary value that captures whether the day is a holiday or not. 0: No holiday, 1: Holiday.}
\item{Ferientyp}{Indicates the type of holidays if it is a holiday, e.g. Sommerferien, Herbstferien.}
\item{Wochentag}{Day of the week, e.g. Montag, Dienstag.}
\item{Wochenende}{Binary value that captures whether the day is a day on the weekend. 0: Mo - Fr, 1: Sa - Su.}
}

An object of class \code{data.frame} with 13140 rows and 12 columns.

An object of class \code{data.frame} with 13140 rows and 12 columns.

An object of class \code{data.frame} with 13140 rows and 12 columns.

An object of class \code{data.frame} with 13176 rows and 12 columns.

An object of class \code{data.frame} with 13140 rows and 12 columns.
}
\source{
\url{https://www.schulferien.org/schweiz/ferien/}
}
\usage{
holidays_2020_long

holidays_2021_long

holidays_2022_long

holidays_2023_long

holidays_2024_long

holidays_2025_long
}
\description{
Datasets in long format containing school holidays (1.- 9. grade) from the
years 2020 - 2025.
}
\keyword{datasets}
