% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays_bar_chart.R
\name{holidays_bar_chart}
\alias{holidays_bar_chart}
\title{Bar chart for holidays datasets}
\usage{
holidays_bar_chart(data)
}
\arguments{
\item{data}{A dataset from this package. This includes holidays_20XX_long,
holidays_20XX_long_major_regions, holidays_20XX_long_switzerland,
holidays_2020_to_2025_long, holidays_2020_to_2025_long_major_regions
and holidays_2020_to_2025_long_switzerland. Not that XX is a
placeholder for 20, 21, 22, 23, 24 or 25.}
}
\value{
An interactive bar plot for the dataset that was passed into the function.
Each bar represents the number of holidays in a year for a given region.
}
\description{
Bar chart for holidays datasets
}
\examples{
df <- data("holidays_2020_to_2025_long_major_regions")
df <- get(df)
holidays_bar_chart(data=df)



}
