% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays_describe.R
\name{holidays_describe}
\alias{holidays_describe}
\title{Holidays dataset variables overview}
\usage{
holidays_describe(data)
}
\arguments{
\item{data}{A dataset from this package. This includes holidays_20XX_long,
holidays_20XX_long_major_regions, holidays_20XX_long_switzerland,
holidays_2020_to_2025_long, holidays_2020_to_2025_long_major_regions
and holidays_2020_to_2025_long_switzerland. Not that XX is a
placeholder for 20, 21, 22, 23, 24 or 25.}
}
\value{
A console output that provides an overview of all variables of
the dataset and unique values they contain.
}
\description{
Holidays dataset variables overview
}
\examples{
df <- data("holidays_2020_long")
df <- get(df)
holidays_describe(data=df)



}
