% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays_heat_map.R
\name{holidays_heat_map}
\alias{holidays_heat_map}
\title{Heat map for holidays datasets}
\usage{
holidays_heat_map(
  data,
  normalize = FALSE,
  holidaytype = c("Sportferien", "Fruehlingsferien", "Sommerferien", "Herbstferien",
    "Weihnachtsferien")
)
}
\arguments{
\item{data}{A dataset from this package. This includes holidays_20XX_long,
holidays_20XX_long_major_regions, holidays_20XX_long_switzerland,
holidays_2020_to_2025_long, holidays_2020_to_2025_long_major_regions
and holidays_2020_to_2025_long_switzerland. Not that XX is a
placeholder for 20, 21, 22, 23, 24 or 25.}

\item{normalize}{Can be set to TRUE or FALSE. If set to TRUE, values for 'Ferientag'
are normalized. This is useful for e.g. holidays_20XX_long_major_regions.
Since not all major regions have the same number of members (e.g. Zentralschweiz: 9,
Genferseeregion: 4). Normalizing the data can make the heat map more meaningful.}

\item{holidaytype}{The type of holidays you want the plot for. Possible values are:
Sportferien, Fruehlingsferien, Sommerferien, Herbstferien, and Weihnachtsferien.
If you don't pass any value here, all types are included.}
}
\value{
An interactive heat map for the dataset that was passed into the function.
The y-axis represents regions ('Kanton', 'Grossregion', or 'Land') whereas
'Datum' is plotted on the x-axis. The heat is represented by 'Ferientag'.
}
\description{
Heat map for holidays datasets
}
\examples{
df <- data("holidays_2020_long")
df <- get(df)
holidays_heat_map(data=df)



}
