% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays_time_series.R
\name{holidays_time_series}
\alias{holidays_time_series}
\title{Time series for holidays datasets}
\usage{
holidays_time_series(data)
}
\arguments{
\item{data}{A dataset from this package. This includes holidays_20XX_long,
holidays_20XX_long_major_regions, holidays_20XX_long_switzerland,
holidays_2020_to_2025_long, holidays_2020_to_2025_long_major_regions
and holidays_2020_to_2025_long_switzerland. Not that XX is a
placeholder for 20, 21, 22, 23, 24 or 25.}
}
\value{
An interactive time series plot for the dataset that was passed into the function.
}
\description{
Time series for holidays datasets
}
\examples{
df <- data("holidays_2020_long")
df <- get(df)
holidays_time_series(data=df)



}
