% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_to_df.R
\name{html_to_df}
\alias{html_to_df}
\title{Converts html table element to df}
\usage{
html_to_df(table)
}
\arguments{
\item{table}{The html table element you wish to convert.}
}
\value{
A data frame.
}
\description{
Converts html table element to df
}
\examples{

\dontrun{
year <- 2022
html <- fetch_html_element(year)
df <- html_to_df(html)
df
}



}
