% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Title}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{String to be split}

\item{split}{Character at which to split}
}
\value{
Returns a list
}
\description{
Title
}
\examples{
x <- "alfa,bravo,charlie,delta"
strsplit1(x, split = ",")
}
