% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays_heat_map.R
\name{holidays_heat_map}
\alias{holidays_heat_map}
\title{Heat map for holidays datasets}
\usage{
holidays_heat_map(data)
}
\arguments{
\item{data}{A dataset from this package. This includes holidays_20XX_long,
holidays_20XX_long_major_regions, holidays_20XX_long_switzerland,
holidays_2020_to_2025_long, holidays_2020_to_2025_long_major_regions
and holidays_2020_to_2025_long_switzerland. Not that XX is a
placeholder for 20, 21, 22, 23, 24 or 25.}
}
\value{
An interactive heat map for the dataset that was passed into the function.
The y-axis represents regions ('Kanton', 'Grossregion', or 'Land') whereas
'Datum' is plotted on the x-axis. The heat is represented by 'Ferientag'.
}
\description{
Heat map for holidays datasets
}
\examples{
df <- data("holidays_2020_long")
df <- get(df)
holidays_heat_map(data=df)



}
